move state augmentation code to dedicated file in `augmentation/`
