+++ openscvx/trajoptproblem.py
@@ -17,7 +17,8 @@ from openscvx.config import (
     DevConfig,
     Config,
 )
-from openscvx.dynamics import Dynamics, build_augmented_dynamics
+from openscvx.dynamics import Dynamics
+from openscvx.augmentation.state_augmentation import build_augmented_dynamics
 from openscvx.augmentation.ctcs_violation import get_g_funcs, CTCSViolation
 from openscvx.augmentation.ctcs import sort_ctcs_constraints
 from openscvx.discretization import get_discretization_solver
